requirements.txt
brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
joblib==1.5.1
numpy==2.2.6
ply==3.11
pycparser==2.22
pysocks==1.7.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.9.0
threadpoolctl==3.6.0
urllib3==2.4.0
wheel==0.45.1
win-inet-pton==1.1.0
zstandard==0.23.0
faiss-cpu